satpambot/bot/requirements.txt
Flask>=3.0
flask-socketio>=5.3.6
python-socketio>=5.11.0
python-engineio>=4.9.1
werkzeug>=2.2
python-dotenv>=1.0
requests>=2.31
discord.py>=2.3
aiohttp>=3.9
aiosqlite>=0.19
pytesseract>=0.3
tldextract>=3.1
Pillow>=10.0
openai>=1.0
psutil>=5.9
hypercorn>=0.14
simplejson>=3.19
httpx>=0.25
idna>=3.4
numpy>=1.23.0
SQLAlchemy>=2.0.0
pytz
eventlet>=0.35.2
greenlet>=3.0
opencv-python-headless>=4.10
tzdata>=2024.1
gunicorn>=21.2

# --- SatpamBot security ---
discord.py
Flask
aiohttp
Pillow
python-dotenv
